// src/app/shared/app/models/Claims/iclaims.ts
export interface IClaims {
  sNo?: number;
  branch?: string;
  claimNo?: string;
  status?: string;
  claimStatusNotes?: string;
  clientID?: number;
  clientName?: string;
  carNo?: string;
  policyCertificateNo?: any; // check it
  chassisNumber?: string;
  shipmentName?: string;
  model?: string;
  insuranceCompany?: string;
  className?: string;
  lineofBusiness?: string;
  policyNo?: string;
  membName?: string;
  claimAmount?: number;
  amountPaid?: number;
  amountRejected?: number;
  amountDeducted?: number;
  underProcessingAmount?: number;
  estimatedValue?: number;
  totalApprovals?: number;
  totalPaidToWorkshop?: number;
  towingCharges?: number;
  dateOfLoss?: Date;
  dateOfReceive?: Date;
  dateOfSubmission?: Date;
  dateOfDeadline?: Date;
  approvalDate?: Date;
  invoiceDate?: Date;
  dateofPayment?: Date;
  insurCompClaimNo?: string;
  notes?: string;
  savedUser?: string;
  savedDate?: Date;
  notifyClient?: boolean;
  notifyInsurer?: boolean;
  identity?: string;
}