// src/app/shared/app/models/CustomerService/icustomer-service.ts
export interface ICustomerService {
	sno?: number;
	endorsType?: string;
	branch?: string;
	status?: string;
	endorsNo?: string;
	changeStatus?: string;
	docSno?: number;
	requestNo?: string;
	type?: string;
	requestStatusNote?: string;
	clientId?: number;
	prospectId?: number;
	clientName?: string;
	policyNo?: string;
	policySerial?: number;
	clientPolicySno?: number;
	insurComp?: string;
	classOfBusiness?: string;
	lineOfBusiness?: string;
	existingPolExpDate?: string;
	csspecialConditions?: string;
	requestDetails?: string;
	netPremium?: number;
	policyFees?: number;
	vatPerc?: number;
	vatValue?: number;
	vatAmount?: number;
	totalPremium?: number;
	savedBy?: string;
	savedDate?: Date;
	updatedBy?: string;
	updatedOn?: Date;
	closedBy?: string;
	closedOn?: Date;
	cancelationReason?: string;
	canceledBy?: string;
	canceledOn?: Date;
	finVerified?: number;
	dateOfReminder?: Date;
	dateOfDeadline?: Date;
	notifyClient?: number;
	notifyInsurer?: number;
	producer?: string;
	number?: string;
	email?: string;
	rejectionReason?: string;
}